Bug Fix, Reset if event handler not attached
A bug was discovered that if the event handler was not attached the
arduino could be reset each time an event came in from the display.
Events should always be handled if they are created, however in the case
the user didn't include the handler in their sketch, it would reset the
arduino.