Modifications to genieBegin, and genieWriteStrU
Removed the separate serial read/write functions and the function array
for them.  Added the genieBegin(Serial&) function to adhere to the API
style guide.  Added a debug serial pointer and an interface to set it
(assignDebugPort).

Fixed the library for genieWriteStrU which is for Unicode string
writing, as the code was incorrect and did not work correctly.

Slight change to the genieArduino_Demo.ino to use
genieGetEventData(&Event) by default in the examples for getting data
from the event. Added some more comments.