Library Updated - NOTE Syntax Change required
This library has been updated to be more in line with the Style Guide.
Everything is now wrapped into a class, so previous functions such as
genieWriteObject() are now genie.WriteObject(). Also the way the Serial
Port is defined for use with the library, is now more standard. You
simply do a normal Serial Begin for the serial port of your choice (and
what your device can handle), and then link that with the genie library.
Please refer to the Demo supplied with this release.
The Demo has also been updated and separates the genie.ReadObject
function so it talks to a different object, rather than the same as what
the Events talk to. Should help explain it a little better.
Like always, any bug reporting is appreciated.