Minor Update - Removed Flush to allow >2 ReadObject
Removed the flush in the library which was designed to clear out old
data which may have been sitting, when a ReadObject is called. This was
however preventing more than 2 ReadObjects to be executed in a row. If
this occurred, the last 2 would be executed only and all previous
ReadObject commands would be ignored/flushed. This is now fixed.