Add Support for 2+ Displays on 1 Arduino + New Demo
Added support for more than 1 display to be connected to a single
Arduino, with more than 1 serial port. An Arduino Mega is a good
candidate. Added a new demo to illustrate how to code for more than 1
display, and lets objects on 1 display talk to objects on the other.
Neat.