Add Support for Inherent Labels
Added functions to support Inherent Labels, which are similar to String objects but allow for the use of External Fonts, not just the default FONT1, FONT2, FONT3. Like strings these can be dynamically changed by the Arduino but with a feature is they can be returned to their default information state as defined in WS4.
The length of an Inherent Label buffer is defined in the properties of the widget in Workshop4, using the Buffer parameter. Max is 255.
Added Demo to illustrate control.